made sure that imagenet and mnist notebooks run locally

+++ imagenet_experiments/load_models.py
@@ -1,15 +1,20 @@
 import os
 import torch
 import torchvision.models as models
+import pathlib
+from pathlib import Path
 
-def clearhub(hubdir = "/home/gridsan/gstepaniants/.cache/torch/hub/checkpoints"):
-    filelist = os.listdir(hubdir)
-    for f in filelist:
-        os.remove(os.path.join(hubdir, f))
+home = str(Path.home())
+
+def clearhub(hubdir = f"{home}/.cache/torch/hub/checkpoints"):
+    if os.path.isdir(hubdir):
+        filelist = os.listdir(hubdir)
+        for f in filelist:
+            os.remove(os.path.join(hubdir, f))
 
 clearhub()
 
-pretrained = False
+pretrained = True
 
 suff = "_pretrained"
